[postgres client] lots of fixes
